[Cypress] Refactor domain selection function.
